drivers: spi: stm32: use codegen driver instantiation
Replace device instances code by codegen solution.

Extend bindings to allow DTS extract for device info.

Correct DTS to be used in SPI driver.

Remove defines from dts.fixup that are now taken from DTS.
