+++ arch/arm/soc/st_stm32/stm32f0/dts.fixup
@@ -14,16 +14,6 @@
 #define CONFIG_UART_STM32_USART_2_NAME		ST_STM32_USART_40004400_LABEL
 #define USART_2_IRQ				ST_STM32_USART_40004400_IRQ_0
 
-#define CONFIG_SPI_1_BASE_ADDRESS		ST_STM32_SPI_FIFO_40013000_BASE_ADDRESS
-#define CONFIG_SPI_1_IRQ_PRI			ST_STM32_SPI_FIFO_40013000_IRQ_0_PRIORITY
-#define CONFIG_SPI_1_NAME			ST_STM32_SPI_FIFO_40013000_LABEL
-#define CONFIG_SPI_1_IRQ			ST_STM32_SPI_FIFO_40013000_IRQ_0
-
-#define CONFIG_SPI_2_BASE_ADDRESS		ST_STM32_SPI_FIFO_40003800_BASE_ADDRESS
-#define CONFIG_SPI_2_IRQ_PRI			ST_STM32_SPI_FIFO_40003800_IRQ_0_PRIORITY
-#define CONFIG_SPI_2_NAME			ST_STM32_SPI_FIFO_40003800_LABEL
-#define CONFIG_SPI_2_IRQ			ST_STM32_SPI_FIFO_40003800_IRQ_0
-
 #define CONFIG_CAN_1_BASE_ADDRESS		ST_STM32_CAN_40006400_BASE_ADDRESS
 #define CONFIG_CAN_1_BUS_SPEED			ST_STM32_CAN_40006400_BUS_SPEED
 #define CONFIG_CAN_1_NAME			ST_STM32_CAN_40006400_LABEL
