+++ arch/arm/soc/st_stm32/stm32f1/dts.fixup
@@ -26,21 +26,6 @@
 #define CONFIG_UART_STM32_PORT_4_NAME           ST_STM32_UART_40004C00_LABEL
 #define PORT_4_IRQ                              ST_STM32_UART_40004C00_IRQ_0
 
-#define CONFIG_SPI_1_BASE_ADDRESS		ST_STM32_SPI_40013000_BASE_ADDRESS
-#define CONFIG_SPI_1_IRQ_PRI			ST_STM32_SPI_40013000_IRQ_0_PRIORITY
-#define CONFIG_SPI_1_NAME			ST_STM32_SPI_40013000_LABEL
-#define CONFIG_SPI_1_IRQ			ST_STM32_SPI_40013000_IRQ_0
-
-#define CONFIG_SPI_2_BASE_ADDRESS		ST_STM32_SPI_40003800_BASE_ADDRESS
-#define CONFIG_SPI_2_IRQ_PRI			ST_STM32_SPI_40003800_IRQ_0_PRIORITY
-#define CONFIG_SPI_2_NAME			ST_STM32_SPI_40003800_LABEL
-#define CONFIG_SPI_2_IRQ			ST_STM32_SPI_40003800_IRQ_0
-
-#define CONFIG_SPI_3_BASE_ADDRESS		ST_STM32_SPI_40003C00_BASE_ADDRESS
-#define CONFIG_SPI_3_IRQ_PRI			ST_STM32_SPI_40003C00_IRQ_0_PRIORITY
-#define CONFIG_SPI_3_NAME			ST_STM32_SPI_40003C00_LABEL
-#define CONFIG_SPI_3_IRQ			ST_STM32_SPI_40003C00_IRQ_0
-
 #define CONFIG_USB_BASE_ADDRESS			ST_STM32_USB_40005C00_BASE_ADDRESS
 #define CONFIG_USB_IRQ				ST_STM32_USB_40005C00_IRQ_USB
 #define CONFIG_USB_IRQ_PRI			ST_STM32_USB_40005C00_IRQ_USB_PRIORITY
