+++ arch/arm/soc/st_stm32/stm32f4/dts.fixup
@@ -26,36 +26,6 @@
 #define CONFIG_UART_STM32_USART_6_IRQ_PRI        ST_STM32_USART_40011400_IRQ_0_PRIORITY
 #define USART_6_IRQ                              ST_STM32_USART_40011400_IRQ_0
 
-#define CONFIG_SPI_1_BASE_ADDRESS               ST_STM32_SPI_40013000_BASE_ADDRESS
-#define CONFIG_SPI_1_IRQ_PRI                    ST_STM32_SPI_40013000_IRQ_0_PRIORITY
-#define CONFIG_SPI_1_NAME                       ST_STM32_SPI_40013000_LABEL
-#define CONFIG_SPI_1_IRQ                        ST_STM32_SPI_40013000_IRQ_0
-
-#define CONFIG_SPI_2_BASE_ADDRESS               ST_STM32_SPI_40003800_BASE_ADDRESS
-#define CONFIG_SPI_2_IRQ_PRI                    ST_STM32_SPI_40003800_IRQ_0_PRIORITY
-#define CONFIG_SPI_2_NAME                       ST_STM32_SPI_40003800_LABEL
-#define CONFIG_SPI_2_IRQ                        ST_STM32_SPI_40003800_IRQ_0
-
-#define CONFIG_SPI_3_BASE_ADDRESS		ST_STM32_SPI_40003C00_BASE_ADDRESS
-#define CONFIG_SPI_3_IRQ_PRI			ST_STM32_SPI_40003C00_IRQ_0_PRIORITY
-#define CONFIG_SPI_3_NAME			ST_STM32_SPI_40003C00_LABEL
-#define CONFIG_SPI_3_IRQ			ST_STM32_SPI_40003C00_IRQ_0
-
-#define CONFIG_SPI_4_BASE_ADDRESS		ST_STM32_SPI_40013400_BASE_ADDRESS
-#define CONFIG_SPI_4_IRQ_PRI			ST_STM32_SPI_40013400_IRQ_0_PRIORITY
-#define CONFIG_SPI_4_NAME			ST_STM32_SPI_40013400_LABEL
-#define CONFIG_SPI_4_IRQ			ST_STM32_SPI_40013400_IRQ_0
-
-#define CONFIG_SPI_5_BASE_ADDRESS		ST_STM32_SPI_40015000_BASE_ADDRESS
-#define CONFIG_SPI_5_IRQ_PRI			ST_STM32_SPI_40015000_IRQ_0_PRIORITY
-#define CONFIG_SPI_5_NAME			ST_STM32_SPI_40015000_LABEL
-#define CONFIG_SPI_5_IRQ			ST_STM32_SPI_40015000_IRQ_0
-
-#define CONFIG_SPI_6_BASE_ADDRESS		ST_STM32_SPI_40015400_BASE_ADDRESS
-#define CONFIG_SPI_6_IRQ_PRI			ST_STM32_SPI_40015400_IRQ_0_PRIORITY
-#define CONFIG_SPI_6_NAME			ST_STM32_SPI_40015400_LABEL
-#define CONFIG_SPI_6_IRQ			ST_STM32_SPI_40015400_IRQ_0
-
 #define FLASH_DEV_BASE_ADDRESS		        ST_STM32F4_FLASH_CONTROLLER_40023C00_BASE_ADDRESS
 #define FLASH_DEV_NAME			        ST_STM32F4_FLASH_CONTROLLER_40023C00_LABEL
