+++ arch/arm/soc/st_stm32/stm32l4/dts.fixup
@@ -44,21 +44,6 @@
 #define CONFIG_RTC_0_NAME                       ST_STM32_RTC_40002800_LABEL
 #define CONFIG_RTC_PRESCALER                    ST_STM32_RTC_40002800_PRESCALER
 
-#define CONFIG_SPI_1_BASE_ADDRESS               ST_STM32_SPI_FIFO_40013000_BASE_ADDRESS
-#define CONFIG_SPI_1_IRQ_PRI                    ST_STM32_SPI_FIFO_40013000_IRQ_0_PRIORITY
-#define CONFIG_SPI_1_NAME                       ST_STM32_SPI_FIFO_40013000_LABEL
-#define CONFIG_SPI_1_IRQ                        ST_STM32_SPI_FIFO_40013000_IRQ_0
-
-#define CONFIG_SPI_2_BASE_ADDRESS		ST_STM32_SPI_FIFO_40003800_BASE_ADDRESS
-#define CONFIG_SPI_2_IRQ_PRI			ST_STM32_SPI_FIFO_40003800_IRQ_0_PRIORITY
-#define CONFIG_SPI_2_NAME			ST_STM32_SPI_FIFO_40003800_LABEL
-#define CONFIG_SPI_2_IRQ			ST_STM32_SPI_FIFO_40003800_IRQ_0
-
-#define CONFIG_SPI_3_BASE_ADDRESS               ST_STM32_SPI_FIFO_40003C00_BASE_ADDRESS
-#define CONFIG_SPI_3_IRQ_PRI                    ST_STM32_SPI_FIFO_40003C00_IRQ_0_PRIORITY
-#define CONFIG_SPI_3_NAME                       ST_STM32_SPI_FIFO_40003C00_LABEL
-#define CONFIG_SPI_3_IRQ                        ST_STM32_SPI_FIFO_40003C00_IRQ_0
-
 #define FLASH_DEV_BASE_ADDRESS		        ST_STM32L4_FLASH_CONTROLLER_40022000_BASE_ADDRESS
 #define FLASH_DEV_NAME			        ST_STM32L4_FLASH_CONTROLLER_40022000_LABEL
