+++ dts/bindings/spi/spi.yaml
@@ -15,6 +15,11 @@ child:
     bus: spi
 
 properties:
+    compatible:
+      type: string
+      category: required
+      description: compatible strings
+      generation: define
     "#address-cells":
       type: int
       category: required
@@ -37,6 +42,9 @@ properties:
       type: compound
       category: optional
       generation: define, use-prop-name
-
-
+    pinctrl-\d+:
+      type: array
+      category: optional
+      description: pinmux information for SCK, MOSI, MISO
+      generation: define
 ...
