+++ dts/bindings/spi/st,stm32-spi-fifo.yaml
@@ -28,6 +28,12 @@ properties:
       generation: define
       category: required
 
+    clocks:
+      type: array
+      category: required
+      description: required clocks
+      generation: define
+
     interrupts:
       type: array
       category: required
